stun fx
so its kinda weird bc currently IN BATTLE ONLY status effects tick down on timers which wouldn't be a bad thing but for some reason i feel like these timers could desync but here's hoping for the best. otherwise the stun fx looks much better than i thought it would honestly im pretty happy about it (and to be clear any status effect could have fx though im really considering swapping to planes but i'll have to do more research in blender)

+++ Assets/scripts/MainGameScript/BuffDebuffSystem/BuffHolder.cs
@@ -3,14 +3,14 @@ using UnityEngine;
 public class BuffHolder 
 {
 
-    public int timeRemaining;
+    public float timeRemaining;
     public int buffId;
     public BuffHolder(int time, int buffId)
     {
         this.timeRemaining = time;
         this.buffId = buffId;   
     }
-    public bool ProgressStatus(int time = 1)
+    public bool ProgressStatus(float time = 1)
     {
         timeRemaining -= time;
         if (timeRemaining <= 0) { return true; }
